+++ bundles/framework/featuredata/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('featuredata', () => Oskari.clazz.create('Oskari.mapframework.bundle.featuredata.FeatureDataBundleInstance'));