+++ ReleaseNotes.md
@@ -23,6 +23,7 @@ Migrated bundles from under `packages` to `bundles` and/or to new bundle-loader 
 - packages/admin/bundle/metrics/bundle.js -> bundles/admin/metrics
 
 - packages/framework/layerlist/bundle.js -> bundles/framework/layerlist
+- packages/framework/featuredata/bundle.js -> bundles/framework/featuredata
 
 - packages/mapping/dimension-change/bundle.js -> bundles/mapping/dimension-change
 - packages/mapping/camera-controls-3d/bundle.js -> bundles/mapping/camera-controls-3d